lightbulb: change led_num to uint16 for ws2812

See merge request ae_group/esp-iot-solution!1226

+++ components/led/lightbulb_driver/drivers/ws2812/ws2812.h
@@ -16,7 +16,7 @@ extern "C" {
  *
  */
 typedef struct {
-    uint8_t led_num;
+    uint16_t led_num;
     gpio_num_t ctrl_io;
 } driver_ws2812_t;
